split web js to v1 and v2
