--- flutter/web/README.md
@@ -1,17 +1,0 @@
-# RustDesk web
-
-## Functions
-
-### Current and planned
-
-- [x] Outgoing
-- [ ] Address book
-- [ ] Force relay
-
-### Unsupported
-
-1. Incoming
-2. LAN
-3. Gpu texture render. We use WebGL instead.
-
-### No plans
