--- flutter/web/js/src/consts.ts
@@ -1,19 +1,0 @@
-export const LOGIN_MSG_DESKTOP_SESSION_NOT_READY = 'Desktop session not ready';
-export const LOGIN_MSG_DESKTOP_XSESSION_FAILED = 'Desktop xsession failed';
-export const LOGIN_MSG_DESKTOP_SESSION_ANOTHER_USER = 'Desktop session another user login';
-export const LOGIN_MSG_DESKTOP_XORG_NOT_FOUND = 'Desktop xorg not found';
-// ls /usr/share/xsessions/
-export const LOGIN_MSG_DESKTOP_NO_DESKTOP = 'Desktop none';
-export const LOGIN_MSG_DESKTOP_SESSION_NOT_READY_PASSWORD_EMPTY =
-    'Desktop session not ready, password empty';
-export const LOGIN_MSG_DESKTOP_SESSION_NOT_READY_PASSWORD_WRONG =
-    'Desktop session not ready, password wrong';
-export const LOGIN_MSG_PASSWORD_EMPTY = 'Empty Password';
-export const LOGIN_MSG_PASSWORD_WRONG = 'Wrong Password';
-export const LOGIN_MSG_2FA_WRONG = 'Wrong 2FA Code';
-export const REQUIRE_2FA = '2FA Required';
-export const LOGIN_MSG_NO_PASSWORD_ACCESS = 'No Password Access';
-export const LOGIN_MSG_OFFLINE = 'Offline';
-export const LOGIN_SCREEN_WAYLAND = 'Wayland login screen is not supported';
-export const SCRAP_X11_REQUIRED = 'x11 expected';
-export const SCRAP_X11_REF_URL = 'https://rustdesk.com/docs/en/manual/linux/#x11-required';
