+++ flutter/web/v1/README.md
@@ -1,0 +1,1 @@
+v1 is not compatible with current Flutter source code.
