+++ flutter/web/v1/js/package.json
@@ -3,19 +3,19 @@
   "version": "1.0.0",
   "scripts": {
     "dev": "vite",
-    "build": "python ./gen_js_from_hbb.py > src/gen_js_from_hbb.ts && python ./ts_proto.py && tsc && vite build",
+    "build": "./gen_js_from_hbb.py > src/gen_js_from_hbb.ts && ./ts_proto.py && tsc && vite build",
     "preview": "vite preview"
   },
   "devDependencies": {
     "typescript": "^4.4.4",
-    "vite": "2.8"
+    "vite": "^2.7.2"
   },
   "dependencies": {
     "fast-sha256": "^1.3.0",
     "libsodium": "^0.7.9",
     "libsodium-wrappers": "^0.7.9",
     "pcm-player": "^0.0.11",
-    "ts-proto": "^1.169.1",
+    "ts-proto": "^1.101.0",
     "wasm-feature-detect": "^1.2.11",
     "zstddec": "^0.0.2"
   }
