+++ flutter/web/v1/js/src/connection.ts
@@ -4,10 +4,9 @@ import * as rendezvous from "./rendezvous.js";
 import { loadVp9 } from "./codec";
 import * as sha256 from "fast-sha256";
 import * as globals from "./globals";
-import * as consts from "./consts";
 import { decompress, mapKey, sleep } from "./common";
 
-export const PORT = 21116;
+const PORT = 21116;
 const HOSTS = [
   "rs-sg.rustdesk.com",
   "rs-cn.rustdesk.com",
@@ -16,8 +15,8 @@ const HOSTS = [
 let HOST = localStorage.getItem("rendezvous-server") || HOSTS[0];
 const SCHEMA = "ws://";
 
-type MsgboxCallback = (type: string, title: string, text: string, link: string) => void;
-type DrawCallback = (display: number, data: Uint8Array) => void;
+type MsgboxCallback = (type: string, title: string, text: string) => void;
+type DrawCallback = (data: Uint8Array) => void;
 //const cursorCanvas = document.createElement("canvas");
 
 export default class Connection {
@@ -67,7 +66,7 @@ export default class Connection {
         try {
           this._password = Uint8Array.from(JSON.parse("[" + p + "]"));
         } catch (e) {
-          console.error('Failed to get password, ' + e);
+          console.error(e);
         }
       }
     }
@@ -171,7 +170,7 @@ export default class Connection {
           pk = undefined;
         }
       } catch (e) {
-        console.error('Failed to verify id pk, ', e);
+        console.error(e);
         pk = undefined;
       }
       if (!pk)
@@ -196,7 +195,7 @@ export default class Connection {
     try {
       signedId = await globals.verify(signedId.id, Uint8Array.from(pk!));
     } catch (e) {
-      console.error('Failed to verify signed id pk, ', e);
+      console.error(e);
       // fall back to non-secure connection in case pk mismatch
       console.error("pk mismatch, fall back to non-secure");
       const public_key = message.PublicKey.fromPartial({});
@@ -243,12 +242,26 @@ export default class Connection {
         this.login();
       } else if (msg?.test_delay) {
         const test_delay = msg?.test_delay;
-        console.log('test delay: ', test_delay);
+        console.log(test_delay);
         if (!test_delay.from_client) {
           this._ws?.sendMessage({ test_delay });
         }
       } else if (msg?.login_response) {
-        this.handleLoginResponse(msg?.login_response);
+        const r = msg?.login_response;
+        if (r.error) {
+          if (r.error == "Wrong Password") {
+            this._password = undefined;
+            this.msgbox(
+              "re-input-password",
+              r.error,
+              "Do you want to enter again?"
+            );
+          } else {
+            this.msgbox("error", "Login Error", r.error);
+          }
+        } else if (r.peer_info) {
+          this.handlePeerInfo(r.peer_info);
+        }
       } else if (msg?.video_frame) {
         this.handleVideoFrame(msg?.video_frame!);
       } else if (msg?.clipboard) {
@@ -261,7 +274,7 @@ export default class Connection {
         try {
           globals.copyToClipboard(new TextDecoder().decode(cb.content));
         } catch (e) {
-          console.error('Failed to copy to clipboard, ', e);
+          console.error(e);
         }
         // globals.pushEvent("clipboard", cb);
       } else if (msg?.cursor_data) {
@@ -305,110 +318,13 @@ export default class Connection {
     }
   }
 
-  handleLoginResponse(response: message.LoginResponse) {
-    const loginErrorMap: Record<string, any> = {
-      [consts.LOGIN_SCREEN_WAYLAND]: {
-        msgtype: "error",
-        title: "Login Error",
-        text: "Login screen using Wayland is not supported",
-        link: "https://rustdesk.com/docs/en/manual/linux/#login-screen",
-        try_again: true,
-      },
-      [consts.LOGIN_MSG_DESKTOP_SESSION_NOT_READY]: {
-        msgtype: "session-login",
-        title: "",
-        text: "",
-        link: "",
-        try_again: true,
-      },
-      [consts.LOGIN_MSG_DESKTOP_XSESSION_FAILED]: {
-        msgtype: "session-re-login",
-        title: "",
-        text: "",
-        link: "",
-        try_again: true,
-      },
-      [consts.LOGIN_MSG_DESKTOP_SESSION_ANOTHER_USER]: {
-        msgtype: "info-nocancel",
-        title: "another_user_login_title_tip",
-        text: "another_user_login_text_tip",
-        link: "",
-        try_again: false,
-      },
-      [consts.LOGIN_MSG_DESKTOP_XORG_NOT_FOUND]: {
-        msgtype: "info-nocancel",
-        title: "xorg_not_found_title_tip",
-        text: "xorg_not_found_text_tip",
-        link: "https://rustdesk.com/docs/en/manual/linux/#login-screen",
-        try_again: true,
-      },
-      [consts.LOGIN_MSG_DESKTOP_NO_DESKTOP]: {
-        msgtype: "info-nocancel",
-        title: "no_desktop_title_tip",
-        text: "no_desktop_text_tip",
-        link: "https://rustdesk.com/docs/en/manual/linux/#login-screen",
-        try_again: true,
-      },
-      [consts.LOGIN_MSG_DESKTOP_SESSION_NOT_READY_PASSWORD_EMPTY]: {
-        msgtype: "session-login-password",
-        title: "",
-        text: "",
-        link: "",
-        try_again: true,
-      },
-      [consts.LOGIN_MSG_DESKTOP_SESSION_NOT_READY_PASSWORD_WRONG]: {
-        msgtype: "session-login-re-password",
-        title: "",
-        text: "",
-        link: "",
-        try_again: true,
-      },
-      [consts.LOGIN_MSG_NO_PASSWORD_ACCESS]: {
-        msgtype: "wait-remote-accept-nook",
-        title: "Prompt",
-        text: "Please wait for the remote side to accept your session request...",
-        link: "",
-        try_again: true,
-      },
-    };
-
-    const err = response.error;
-    if (err) {
-      if (err == consts.LOGIN_MSG_PASSWORD_EMPTY) {
-        this._password = undefined;
-        this.msgbox("input-password", "Password Required", "", "");
-      }
-      if (err == consts.LOGIN_MSG_PASSWORD_WRONG) {
-        this._password = undefined;
-        this.msgbox(
-          "re-input-password",
-          err,
-          "Do you want to enter again?"
-        );
-      } else if (err == consts.LOGIN_MSG_2FA_WRONG || err == consts.REQUIRE_2FA) {
-        this.msgbox("input-2fa", err, "");
-      } else if (err in loginErrorMap) {
-        const m = loginErrorMap[err];
-        this.msgbox(m.msgtype, m.title, m.text, m.link);
-      } else {
-        if (err.includes(consts.SCRAP_X11_REQUIRED)) {
-          this.msgbox("error", "Login Error", err, consts.SCRAP_X11_REF_URL);
-        } else {
-          this.msgbox("error", "Login Error", err);
-        }
-      }
-    } else if (response.peer_info) {
-      this.handlePeerInfo(response.peer_info);
-    }
+  msgbox(type_: string, title: string, text: string) {
+    this._msgbox?.(type_, title, text);
   }
 
-  msgbox(type_: string, title: string, text: string, link: string = '') {
-    this._msgbox?.(type_, title, text, link);
-  }
-
-  draw(display: number, frame: any) {
-    this._draw?.(display, frame);
-    globals.draw(display, frame);
+  draw(frame: any) {
+    this._draw?.(frame);
+    globals.draw(frame);
   }
 
   close() {
@@ -431,36 +347,23 @@ export default class Connection {
     this._draw = callback;
   }
 
-  login(info?: {
-    os_login?: message.OSLogin,
-    password?: Uint8Array
-  }) {
-    if (info?.password) {
+  login(password: string | undefined = undefined) {
+    if (password) {
       const salt = this._hash?.salt;
-      let p = hash([info.password, salt!]);
+      let p = hash([password, salt!]);
       this._password = p;
       const challenge = this._hash?.challenge;
       p = hash([p, challenge!]);
       this.msgbox("connecting", "Connecting...", "Logging in...");
-      this._sendLoginMessage({ os_login: info.os_login, password: p });
+      this._sendLoginMessage(p);
     } else {
       let p = this._password;
       if (p) {
         const challenge = this._hash?.challenge;
         p = hash([p, challenge!]);
       }
-      this._sendLoginMessage({ os_login: info?.os_login, password: p });
+      this._sendLoginMessage(p);
     }
-  }
-
-  changePreferCodec() {
-    const supported_decoding = message.SupportedDecoding.fromPartial({
-      ability_vp9: 1,
-      ability_h264: 1,
-    });
-    const option = message.OptionMessage.fromPartial({ supported_decoding });
-    const misc = message.Misc.fromPartial({ option });
-    this._ws?.sendMessage({ misc });
   }
 
   async reconnect() {
@@ -468,18 +371,14 @@ export default class Connection {
     await this.start(this._id);
   }
 
-  _sendLoginMessage(login: {
-    os_login?: message.OSLogin,
-    password?: Uint8Array,
-  }) {
+  _sendLoginMessage(password: Uint8Array | undefined = undefined) {
     const login_request = message.LoginRequest.fromPartial({
       username: this._id!,
       my_id: "web", // to-do
       my_name: "web", // to-do
-      password: login.password,
+      password,
       option: this.getOptionMessage(),
       video_ack_required: true,
-      os_login: login.os_login,
     });
     this._ws?.sendMessage({ login_request });
   }
@@ -536,7 +435,7 @@ export default class Connection {
           i++;
           if (i == n) this.sendVideoReceived();
           if (ok && dec.frameBuffer && n == i) {
-            this.draw(vf.display, dec.frameBuffer);
+            this.draw(dec.frameBuffer);
             const now = new Date().getTime();
             var elapsed = now - tm;
             this._videoTestSpeed[1] += elapsed;
@@ -544,9 +443,9 @@ export default class Connection {
             if (this._videoTestSpeed[0] >= 30) {
               console.log(
                 "video decoder: " +
-                parseInt(
-                  "" + this._videoTestSpeed[1] / this._videoTestSpeed[0]
-                )
+                  parseInt(
+                    "" + this._videoTestSpeed[1] / this._videoTestSpeed[0]
+                  )
               );
               this._videoTestSpeed = [0, 0];
             }
@@ -557,17 +456,8 @@ export default class Connection {
   }
 
   handlePeerInfo(pi: message.PeerInfo) {
-    localStorage.setItem('last_remote_id', this._id);
     this._peerInfo = pi;
-    if (pi.current_display > pi.displays.length) {
-      pi.current_display = 0;
-    }
-    if (globals.getVersionNumber(pi.version) < globals.getVersionNumber("1.1.10")) {
-      this.setPermission("restart", false);
-    }
     if (pi.displays.length == 0) {
-      this.setOption("info", pi);
-      globals.pushEvent("update_privacy_mode", {});
       this.msgbox("error", "Remote Error", "No Display");
       return;
     }
@@ -577,7 +467,6 @@ export default class Connection {
     if (p) this.inputOsPassword(p);
     const username = this.getOption("info")?.username;
     if (username && !pi.username) pi.username = username;
-    globals.pushEvent("update_privacy_mode", {});
     this.setOption("info", pi);
     if (this.getRemember()) {
       if (this._password?.length) {
@@ -590,10 +479,6 @@ export default class Connection {
     } else {
       this.setOption("password", undefined);
     }
-  }
-
-  setPermission(name: string, value: Boolean) {
-    globals.pushEvent("permission", { [name]: value });
   }
 
   shouldAutoLogin(): string {
@@ -631,7 +516,7 @@ export default class Connection {
         default:
           return;
       }
-      this.setPermission(name, p.enabled);
+      globals.pushEvent("permission", { [name]: p.enabled });
     } else if (misc.switch_display) {
       this.loadVideoDecoder();
       globals.pushEvent("switch_display", misc.switch_display);
@@ -652,27 +537,7 @@ export default class Connection {
   }
 
   getOption(name: string): any {
-    return this._options[name] ?? globals.getUserDefaultOption(name);
-  }
-
-  getToggleOption(name: string): Boolean {
-    // TODO: more default settings
-    const defaultToggleTrue = [
-      'show-remote-cursor',
-      'privacy-mode',
-      'enable-file-copy-paste',
-      'allow_swap_key',
-    ];
-    return this._options[name] || (defaultToggleTrue.includes(name) ? true : false);
-  }
-
-  // TODO:
-  getStatus(): String {
-    return JSON.stringify({ status_num: 10 });
-  }
-
-  // TODO:
-  checkConnStatus() {
+    return this._options[name];
   }
 
   setOption(name: string, value: any) {
@@ -727,68 +592,15 @@ export default class Connection {
     this._ws?.sendMessage({ key_event });
   }
 
-  restart() {
-    const misc = message.Misc.fromPartial({});
-    misc.restart_remote_device = true;
-    this._ws?.sendMessage({ misc });
-  }
-
   inputString(seq: string) {
     const key_event = message.KeyEvent.fromPartial({ seq });
     this._ws?.sendMessage({ key_event });
   }
 
-  send2fa(code: string) {
-    const auth_2fa = message.Auth2FA.fromPartial({ code });
-    this._ws?.sendMessage({ auth_2fa });
-  }
-
-  _captureDisplays({ add, sub, set }: {
-    add?: number[], sub?: number[], set?: number[]
-  }) {
-    const capture_displays = message.CaptureDisplays.fromPartial({ add, sub, set });
-    const misc = message.Misc.fromPartial({ capture_displays });
+  switchDisplay(display: number) {
+    const switch_display = message.SwitchDisplay.fromPartial({ display });
+    const misc = message.Misc.fromPartial({ switch_display });
     this._ws?.sendMessage({ misc });
-  }
-
-  switchDisplay(v: string) {
-    try {
-      const obj = JSON.parse(v);
-      const value = obj.value;
-      const isDesktop = obj.isDesktop;
-      if (value.length == 1) {
-        const switch_display = message.SwitchDisplay.fromPartial({ display: value[0] });
-        const misc = message.Misc.fromPartial({ switch_display });
-        this._ws?.sendMessage({ misc });
-
-        if (!isDesktop) {
-          this._captureDisplays({ set: value });
-        } else {
-          // If support merging images, check_remove_unused_displays() in ui_session_interface.rs
-        }
-      } else {
-        this._captureDisplays({ set: value });
-      }
-    }
-    catch (e) {
-      console.log('Failed to switch display, invalid param "' + v + '"');
-    }
-  }
-
-  elevateWithLogon(value: string) {
-    try {
-      const obj = JSON.parse(value);
-      const logon = message.ElevationRequestWithLogon.fromPartial({
-        username: obj.username,
-        password: obj.password
-      });
-      const elevation_request = message.ElevationRequest.fromPartial({ logon });
-      const misc = message.Misc.fromPartial({ elevation_request });
-      this._ws?.sendMessage({ misc });
-    }
-    catch (e) {
-      console.log('Failed to elevate with logon, invalid param "' + value + '"');
-    }
   }
 
   async inputOsPassword(seq: string) {
@@ -839,52 +651,6 @@ export default class Connection {
   }
 
   toggleOption(name: string) {
-
-    //   } else if name == "block-input" {
-    //     option.block_input = BoolOption::Yes.into();
-    // } else if name == "unblock-input" {
-    //     option.block_input = BoolOption::No.into();
-    // } else if name == "show-quality-monitor" {
-    //     config.show_quality_monitor.v = !config.show_quality_monitor.v;
-    // } else if name == "allow_swap_key" {
-    //     config.allow_swap_key.v = !config.allow_swap_key.v;
-    // } else if name == "view-only" {
-    //     config.view_only.v = !config.view_only.v;
-    //     let f = |b: bool| {
-    //         if b {
-    //             BoolOption::Yes.into()
-    //         } else {
-    //             BoolOption::No.into()
-    //         }
-    //     };
-    //     if config.view_only.v {
-    //         option.disable_keyboard = f(true);
-    //         option.disable_clipboard = f(true);
-    //         option.show_remote_cursor = f(true);
-    //         option.enable_file_transfer = f(false);
-    //         option.lock_after_session_end = f(false);
-    //     } else {
-    //         option.disable_keyboard = f(false);
-    //         option.disable_clipboard = f(self.get_toggle_option("disable-clipboard"));
-    //         option.show_remote_cursor = f(self.get_toggle_option("show-remote-cursor"));
-    //         option.enable_file_transfer = f(self.config.enable_file_transfer.v);
-    //         option.lock_after_session_end = f(self.config.lock_after_session_end.v);
-    //     }
-    // } else {
-    //     let is_set = self
-    //         .options
-    //         .get(&name)
-    //         .map(|o| !o.is_empty())
-    //         .unwrap_or(false);
-    //     if is_set {
-    //         self.config.options.remove(&name);
-    //     } else {
-    //         self.config.options.insert(name, "Y".to_owned());
-    //     }
-    //     self.config.store(&self.id);
-    //     return None;
-    // }
-
     const v = !this._options[name];
     const option = message.OptionMessage.fromPartial({});
     const v2 = v
@@ -906,62 +672,18 @@ export default class Connection {
       case "privacy-mode":
         option.privacy_mode = v2;
         break;
-      case "enable-file-copy-paste":
-        option.enable_file_transfer = v2;
-        break;
       case "block-input":
         option.block_input = message.OptionMessage_BoolOption.Yes;
         break;
       case "unblock-input":
         option.block_input = message.OptionMessage_BoolOption.No;
         break;
-      case "show-quality-monitor":
-      case "allow-swap-key":
-        break;
-      case "view-only":
-        if (v) {
-          option.disable_keyboard = message.OptionMessage_BoolOption.Yes;
-          option.disable_clipboard = message.OptionMessage_BoolOption.Yes;
-          option.show_remote_cursor = message.OptionMessage_BoolOption.Yes;
-          option.enable_file_transfer = message.OptionMessage_BoolOption.No;
-          option.lock_after_session_end = message.OptionMessage_BoolOption.No;
-        } else {
-          option.disable_keyboard = message.OptionMessage_BoolOption.No;
-          option.disable_clipboard = this.getToggleOption("disable-clipboard")
-            ? message.OptionMessage_BoolOption.Yes
-            : message.OptionMessage_BoolOption.No;
-          option.show_remote_cursor = this.getToggleOption("show-remote-cursor")
-            ? message.OptionMessage_BoolOption.Yes
-            : message.OptionMessage_BoolOption.No;
-          option.enable_file_transfer = this.getToggleOption("enable-file-copy-paste")
-            ? message.OptionMessage_BoolOption.Yes
-            : message.OptionMessage_BoolOption.No;
-          option.lock_after_session_end = this.getToggleOption("lock-after-session-end")
-            ? message.OptionMessage_BoolOption.Yes
-            : message.OptionMessage_BoolOption.No;
-        }
-        break;
       default:
-        this.setOption(name, this._options[name] ? undefined : "Y");
         return;
     }
     if (name.indexOf("block-input") < 0) this.setOption(name, v);
     const misc = message.Misc.fromPartial({ option });
     this._ws?.sendMessage({ misc });
-  }
-
-  togglePrivacyMode(value: string) {
-    try {
-      const obj = JSON.parse(value);
-      const toggle_privacy_mode = message.TogglePrivacyMode.fromPartial({
-        impl_key: obj.impl_key,
-        on: obj.on,
-      });
-      const misc = message.Misc.fromPartial({ toggle_privacy_mode });
-      this._ws?.sendMessage({ misc });
-    } catch (e) {
-      console.log('Failed to toggle privacy mode, invalid param "' + value + '"')
-    }
   }
 
   getImageQuality() {
@@ -998,7 +720,7 @@ export default class Connection {
     loadVp9((decoder: any) => {
       this._videoDecoder = decoder;
       console.log("vp9 loaded");
-      console.log('The decoder: ', decoder);
+      console.log(decoder);
     });
   }
 }
