+++ flutter/web/v1/js/vite.config.js
